Første utkast til engelsk ferdig.
