+++ src/artikler/sok-sosialhjelp/SokSosialhjelp.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import {detekterSprak, Sprak} from "../../utils/sprakUtils";
 import {Oversettelser} from "../../komponenter/oversettelser/Oversettelser";
-import SokSosialhjelpNynorsk from "./SokSosialhjelpNynorsk";
 import SokSosialhjelpBokmalKrise from "./SokSosialhjelpBokmalKrise";
 import SokSosialhjelpEngelskKrise from "./SokSosialhjelpEngelskKrise";
 import "./sokSosialhjelp.less";
+import SokSosialhjelpNynorskKrise from "./SokSosialhjelpNynorskKrise";
 
 const SokSosialhjelp: React.FC = () => {
     const valgtSprak: string = detekterSprak();
@@ -13,7 +13,7 @@ const SokSosialhjelp: React.FC = () => {
             sprak={[Sprak.NORSK_BOKMAL, Sprak.NYNORSK, Sprak.ENGELSK]}
         >
             {valgtSprak === Sprak.NORSK_BOKMAL && <SokSosialhjelpBokmalKrise />}
-            {valgtSprak === Sprak.NYNORSK && <SokSosialhjelpNynorsk />}
+            {valgtSprak === Sprak.NYNORSK && <SokSosialhjelpNynorskKrise />}
             {valgtSprak === Sprak.ENGELSK && <SokSosialhjelpEngelskKrise />}
         </Oversettelser>
     );
